% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anova_likelihood_ratio_test}
\alias{anova_likelihood_ratio_test}
\title{Likelihood ratio test for comparing two survey indices.}
\usage{
anova_likelihood_ratio_test(m1, m2)
}
\arguments{
\item{m1}{model 1}

\item{m2}{model 2}
}
\value{
A p-value.
}
\description{
Previously called anova_SI. TOLDEO.
}
\examples{
# Simple example
dat <- datasets::airquality
m1 <- mgcv::gam(Ozone ~ s(Solar.R), data = dat)
m1
m2 <- mgcv::gam(Ozone ~ Wind + s(Solar.R), data = dat)
m2
# anova_likelihood_ratio_test(m1 = m1, m2 = m2)
}
