% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_distance}
\alias{calc_distance}
\title{Get distance (km) between two locations
previously called gcd.hf}
\usage{
calc_distance(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{Numeric Decimal degrees. Longitude of the first location.}

\item{lat1}{Numeric Decimal degrees. Latitude of the first location.}

\item{long2}{Numeric Decimal degrees. Longitude of the second location.}

\item{lat2}{Numeric Decimal degrees. Latitude of the second location.}
}
\value{
Distance between locations in km.
}
\description{
Get distance (km) between two locations
previously called gcd.hf
}
\examples{
calc_distance(long1 = 170, lat1 = 62, long2 = 170.1, lat2 = 62.1)
calc_distance(long1 = -170, lat1 = -62, long2 = -170.1, lat2 = -62.1)
calc_distance(long1 = 170, lat1 = -62, long2 = 170.1, lat2 = -62.1)
calc_distance(long1 = 50, lat1 = 50, long2 = 50.1, lat2 = 50.1)
}
