% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{consistency_internal}
\alias{consistency_internal}
\title{Calculate internal consistency of a survey index.}
\usage{
consistency_internal(tt, print_plot = FALSE)
}
\arguments{
\item{tt}{A matrix with survey indices (rows=years, cols=ages)}

\item{print_plot}{Plot it?}
}
\value{
a vector of consistencies
}
\description{
Calculate internal consistency of a survey index.
}
\details{
Previously called internalCons.
}
