% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_crs}
\alias{convert_crs}
\title{Change CRS of coordinates}
\usage{
convert_crs(x, y, crs_in = "+proj=longlat +datum=WGS84", crs_out = "EPSG:3338")
}
\arguments{
\item{x}{Decimal degrees longititude.}

\item{y}{Decimal degrees latitude.}

\item{crs_in}{Default = "+proj=longlat +datum=WGS84".}

\item{crs_out}{Default = "EPSG:3338".}
}
\value{
data.frame with 3 columns: ID, Longitude, and Latitude
}
\description{
Change CRS of coordinates
}
\examples{
convert_crs(x = 170, y = 62)
dat <- surveyIndex::noaa_afsc_public_foss[,c("longitude_dd", "latitude_dd")]
head(dat)
ll <- surveyIndex::convert_crs( # project data
   x = dat$longitude_dd,
   y = dat$latitude_dd,
   crs_in = "+proj=longlat +datum=WGS84",
   crs_out = "EPSG:3338")
head(ll)
}
