% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_deg_rad}
\alias{convert_deg_rad}
\title{Convert decimal degrees to radians
Previously called deg2rad.}
\usage{
convert_deg_rad(deg)
}
\arguments{
\item{deg}{Numeric Decimal degrees of a latitude or longitude.}
}
\value{
The value of degrees in radians.
}
\description{
Convert decimal degrees to radians
Previously called deg2rad.
}
\examples{
convert_deg_rad(170.1)
convert_deg_rad(-60.1)
}
