% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_datrasraw}
\alias{get_datrasraw}
\title{Data frame with one row per haul to DATRASraw alike object}
\usage{
get_datrasraw(x)
}
\arguments{
\item{x}{Data.frame. Zero filled catch data.}
}
\description{
Data frame with one row per haul to DATRASraw alike object
}
\examples{
dat <- data.frame(
          species = 1:2,
          Year = 2016:2020,
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))
 ds <- split(dat,dat$species)
 lapply(ds, get_datrasraw)

}
