% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_effect}
\alias{get_effect}
\title{Calculate confidence intervals for a named parameter in a survey index model.}
\usage{
get_effect(x, dat, parName = "Gear", cutOff, nboot = 1000, pOnly = FALSE)
}
\arguments{
\item{x}{survey index}

\item{dat}{DATRASraw object}

\item{parName}{name of the parameter, e.g. "Gear"}

\item{cutOff}{see getsdmgamindex()}

\item{nboot}{see getsdmgamindex()}

\item{pOnly}{only calculate for positive part of model, defaults to FALSE.}
}
\value{
list of estimates + ci bounds for each age group.
}
\description{
Calculate confidence intervals for a named parameter in a survey index model.
}
