% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_prediction_grid}
\alias{get_prediction_grid}
\title{Prediction a grid for a specific year}
\usage{
get_prediction_grid(year, x, subsel = NULL, varsbyyr = NULL, vars = NULL)
}
\arguments{
\item{year}{Numeric string. The years the data is available for.}

\item{x}{Data.frame. The data.frame of the covariate data going into the model.}

\item{subsel}{Default = NULL.}

\item{varsbyyr}{Character string. The name of the variables that vary by year.}

\item{vars}{Character string. The name of the variables that do not vary by year.}
}
\value{
list of data by year
}
\description{
Prediction a grid for a specific year
}
\examples{
dat <- data.frame(
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB2019 = rnorm(n = 10, mean = 20, sd = 10),
          covB2020 = rnorm(n = 10, mean = 20, sd = 10),
          covB2021 = rnorm(n = 10, mean = 20, sd = 10),
          covB2022 = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))

vars <- "covA"
varsbyyr <- "covB"
YEARS <- 2019:2022
pg <- lapply(YEARS,
       FUN = get_prediction_grid,
       x = dat,
       vars = vars,
       varsbyyr = varsbyyr)
names(pg) <- YEARS
pg
}
