% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surveyidx}
\alias{get_surveyidx}
\title{Calculate survey indices by age.}
\usage{
get_surveyidx(
  x,
  ages,
  myids,
  kvecP = rep(12 * 12, length(ages)),
  kvecZ = rep(8 * 8, length(ages)),
  gamma = 1.4,
  cutOff = 1,
  fam = "Gamma",
  useBIC = FALSE,
  nBoot = 1000,
  mc.cores = 1,
  method = "ML",
  predD = NULL,
  modelZ = NULL,
  modelP = NULL,
  knotsP = NULL,
  knotsZ = NULL,
  predfix = NULL,
  linkZ = "logit",
  CIlevel = 0.95,
  ...
)
}
\arguments{
\item{x}{DATRASraw object}

\item{ages}{vector of ages}

\item{myids}{haul.ids for grid}

\item{kvecP}{vector with spatial smoother max. basis dimension for each age group, strictly positive part of model}

\item{kvecZ}{vector with spatial smoother max. basis dimension for each age group, presence/absence part of model (ignored for Tweedie models)}

\item{gamma}{model degress of freedom inflation factor (see 'gamma' argument to mgcv::gam() )}

\item{cutOff}{treat observations below this value as zero}

\item{fam}{distribution, either "Gamma","LogNormal", or "Tweedie".}

\item{useBIC}{use BIC for smoothness selection (overrides 'gamma' argument)}

\item{nBoot}{number of bootstrap samples used for calculating index confidence intervals}

\item{mc.cores}{number of cores for parallel processing}

\item{method}{smoothness selection method used by 'gam'}

\item{predD}{optional DATRASraw object or data.frame (or named list with such objects, one for each year with names(predD) being the years) , defaults to NULL. If not null this is used as grid.}

\item{modelZ}{vector of model formulae for presence/absence part, one pr. age group (ignored for Tweedie models)}

\item{modelP}{vector of model formulae for strictly positive repsonses, one pr. age group}

\item{knotsP}{optional list of knots to gam, strictly positive repsonses}

\item{knotsZ}{optional list of knots to gam, presence/absence}

\item{predfix}{optional named list of extra variables (besides Gear, HaulDur, Ship, and TimeShotHour),  that should be fixed during prediction step (standardized)}

\item{linkZ}{link function for the grDevices::dev.new part of the model, default: "logit" (not used for Tweedie models).}

\item{CIlevel}{Confidence interval level, defaults to 0.95.}

\item{...}{Optional extra arguments to "gam"}
}
\value{
A survey index (list)
}
\description{
Calculate survey indices by age.
}
\details{
This is based on the methods described in
Berg et al. (2014): "Evaluation of alternative age-based methods for estimating relative abundance from survey data in relation to assessment models",
Fisheries Research 151(2014) 91-99.
}
