% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surveyidx_aic}
\alias{get_surveyidx_aic}
\title{Akaike Information Criterion (or BIC) for survey index models}
\usage{
get_surveyidx_aic(x, BIC = FALSE)
}
\arguments{
\item{x}{survey index as return from get_surveyidx}

\item{BIC}{if TRUE compute BIC instead of AIC}
}
\value{
numeric value
}
\description{
Akaike Information Criterion (or BIC) for survey index models
}
\details{
previously named AIC.surveyIdx.
}
