% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surveyidx_resid}
\alias{get_surveyidx_resid}
\title{Randomized quantile residuals for class 'sdmgamindex'}
\usage{
get_surveyidx_resid(x, a = 1)
}
\arguments{
\item{x}{An object of type 'sdmgamindex' as created by 'get_surveyidx'}

\item{a}{age group}
}
\value{
A vector of residuals, which should be iid standard normal distributed
}
\description{
Randomized quantile residuals for class 'sdmgamindex'
}
\details{
Previously named residuals.surveyIdx.
}
