% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surveyidx_stratmean}
\alias{get_surveyidx_stratmean}
\title{Survey index using the stratified mean method using ICES statistical rectangles as strata.}
\usage{
get_surveyidx_stratmean(x, ageCols, doLog = FALSE)
}
\arguments{
\item{x}{DATRASraw object. Must contain a matrix: x[[2]]$Nage.}

\item{ageCols}{which columns of the Nage matrix should be included?}

\item{doLog}{log-transform?}
}
\value{
a matrix with survey indices
}
\description{
Survey index using the stratified mean method using ICES statistical rectangles as strata.
}
\details{
Previously named get_surveyidxStratMean.
}
