% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{pred_grid_ebs}
\alias{pred_grid_ebs}
\title{EBS Prediction Grid}
\format{
A data frame with 36690 observations on the following 3 variables.
\describe{
  \item{\code{lon}}{Numeric; Longitude (one hundred thousandth of a decimal degree).}
  \item{\code{lat}}{Numeric; Latitude (one hundred thousandth of a decimal degree).}
  \item{\code{Shape_Area}}{Numeric; The area that this location represents. }
  }
}
\source{
https://github.com/James-Thorson-NOAA/VAST
}
\usage{
data("pred_grid_ebs")
}
\description{
EBS Prediction Grid
}
\details{
DETAILS
}
\examples{
data(pred_grid_ebs)
}
\author{
Emily Markowitz (emily.markowitz AT noaa.gov) and James Thorson (james.thorson AT noaa.gov)
}
\keyword{bering}
\keyword{eastern}
\keyword{grid}
\keyword{prediciton}
\keyword{sea}
