% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_surveyidx_grid}
\alias{plot_surveyidx_grid}
\title{Plot a surveyIndexGrid}
\usage{
plot_surveyidx_grid(grid, pch = 1, gridCol = "lightgrey")
}
\arguments{
\item{grid}{a surveyIndexGrid (as created by the "get_grid" function)}

\item{pch}{Inherited from base::plot(). plotting ‘character’, i.e., symbol to use.}

\item{gridCol}{Color of grid plot output.}
}
\value{
nothing
}
\description{
Plot a surveyIndexGrid
}
\examples{
# TOLEDO
}
